Initial windows support. Black screen, but gameplay is running. Mouse clicks working

+++ Engine/Misc/WoflMath.h
@@ -108,7 +108,7 @@ public:
 	
 	inline float Length()
 	{
-		return sqrt(LengthSquared());
+		return sqrtf(LengthSquared());
 	}
 	
 	/** Safe normal */
